Push the dispatcher pin stamp straight to main instead of opening a pull request
The stamp pull request asked a human to approve a release that the
cli-release environment approval had already approved, carrying content
that is fully machine-verified against the release attestations. In
practice it was merged unread, and each one left four action_required
pull_request runs behind because bot-authored pull requests wait for
workflow approval while the automation dispatched the same checks itself.

- Replace open-dispatcher-pin-pr with push-dispatcher-pin: stamp from the
  main tip, verify, commit, and fast-forward push to main; no gh calls
- Mint a GitHub App token in post-publish (ruleset bypass actor) because
  GITHUB_TOKEN cannot bypass the main pull request requirement; drop the
  pull-requests and actions write permissions the pull request needed
- Remove the now unused DispatchPullRequestChecksForHead entry point
- Point the freshness guard message at re-running post-publish
- Record the decision, alternatives, and reversal condition in ADR 0005
  and document the one-time App, variable, secret, and ruleset setup
